Something

+++ Shaders/postfx.frag.glsl
@@ -1,5 +1,7 @@
+#version 100
+
 #ifdef GL_ES
-precision highp float;
+precision mediump float;
 #endif
 
 //uniform vec3 color;
